fixup: built-in functions can't be exotic any more
PR #2969 changed the definition of "ordinary function"
to include all built-in functions (however implemented).

So change "an exotic built-in function object" to
"a built-in function object that is not an ECMAScript function object".

Yes, this is somewhat ugly. I'm hoping it's temporary
until the spec stops caring about how a built-in is implemented.